Fix the Windows desktop build and startup
The desktop (clang-cl) build had not been exercised since the Android work
and neither built nor ran Bloodborne.

Build:
- Keep <windows.h> out of widely included headers. The kernel sync
  primitives (TimedMutex, Semaphore) call Win32 through thin wrappers in
  sync/win32_wait.{h,cpp}; debug_state.h, thread.h and exception.h spell
  DWORD/PCONTEXT out instead of including the SDK. Only .cpp files that
  call Win32 include it now, so its macros (DELETE, WAIT_TIMEOUT, FAR,
  MemoryBarrier) no longer leak into unrelated code.
- #undef MemoryBarrier in vk_command_recorder.h, as vk_rasterizer.cpp does.
- Add the missing #pragma once to five headers.
- Wrap the FFmpeg include of videodec/video_utils.cpp in extern "C".
- Define VK_ENABLE_BETA_EXTENSIONS target-wide: a file that includes
  vulkan_core.h before vk_common.h otherwise hides the beta enums from
  vulkan.hpp later in the same translation unit.
- Build guest_vr_sensor on every platform; hmd/move/vr_tracker and the
  presenter use it everywhere, so the desktop link failed.
- Bump Foundation for the clang-cl __PRETTY_FUNCTION__ fix (magic_enum
  reflection) and the FAR/NEAR shield in TFrustum.

Runtime:
- The linker only accepted the libc/Fios DTPMOD64 self-module marker on the
  guest memory backend and threw on the desktop while loading libSceFios2.
  The marker is a property of the guest binary, so accept it everywhere.
- Drop the speculative storage usage of an image when the driver rejects
  it for the format (BC6H on desktop drivers) instead of asserting; views
  follow the backing's actual usage.
- The Windows vectored handler reported every exception it saw first and
  shut the emulator down, including first-chance C++/WIL exceptions that
  system DLLs (the orientation-sensor API on handhelds) throw and catch
  themselves. Software-raised exceptions now go to their own handlers; a
  top-level filter reports the ones nobody handles, naming the C++ type and
  what() for MSVC-ABI exceptions.

+++ src/core/thread.h
@@ -11,7 +11,7 @@ struct PthreadAttr;
 
 namespace Core {
 #ifdef WIN32
-using ThreadFunc = DWORD (*)(void*);
+using ThreadFunc = unsigned long (*)(void*); // DWORD; <windows.h> stays in thread.cpp
 #else
 using ThreadFunc = void* (*)(void*);
 #endif
